@@ -1,5 +1,5 @@
 class FriendshipsController < ApplicationController
-  before_filter :require_login
+  before_action :require_login
 
   def create
     @friendship = current_user.friendships.create!(:friend_id => params[:friend_id])
@@ -10,5 +10,10 @@ class FriendshipsController < ApplicationController
     @friendship = current_user.friendships.find(params[:id])
     @friendship.destroy
     redirect_to :back, notice: "Removed friendship."
+  end
+
+  private
+  def comment_params
+    params.require(:friendship).permit(:friend_id)
   end
 end
